visual_dsa/requirements.txt
streamlit
groq
python-dotenv
manim
